Change calculate_pi to be split being the Pi driver and the calculate_pi package
