+++ example/calculate_pi/calculate_pi.hpp
@@ -10,39 +10,25 @@
 // license in this material to reproduce, prepare derivative works, distribute copies to
 // the public, perform publicly and display publicly, and to permit others to do so.
 //========================================================================================
-#ifndef EXAMPLE_CALCULATE_PI_PI_HPP_
-#define EXAMPLE_CALCULATE_PI_PI_HPP_
+#ifndef EXAMPLE_CALCULATE_PI_CALCULATE_PI_HPP_
+#define EXAMPLE_CALCULATE_PI_CALCULATE_PI_HPP_
 
 // Standard Includes
 #include <memory>
 
 // Parthenon Includes
-#include <parthenon/driver.hpp>
 #include <parthenon/package.hpp>
 
 namespace calculate_pi {
 using namespace parthenon::package::prelude;
-using namespace parthenon::driver::prelude;
 
-/**
- * @brief Constructs a driver which estimates PI using AMR.
- */
-class CalculatePi : public Driver {
- public:
-  CalculatePi(ParameterInput *pin, Mesh *pm, Outputs *pout) : Driver(pin, pm, pout) {}
-
-  /// MakeTaskList isn't a virtual routine on `Driver`, but each driver is expected to
-  /// implement it.
-  TaskList MakeTaskList(MeshBlock *pmb);
-
-  /// `Execute` cylces until simulation completion.
-  DriverStatus Execute() override;
-};
-
+// Package Callbacks
 void SetInOrOut(Container<Real> &rc);
 parthenon::AmrTag CheckRefinement(Container<Real> &rc);
 std::shared_ptr<StateDescriptor> Initialize(ParameterInput *pin);
 
+// Task Implementations
+parthenon::TaskStatus ComputeArea(parthenon::MeshBlock *pmb);
 } // namespace calculate_pi
 
-#endif // EXAMPLE_CALCULATE_PI_PI_HPP_
+#endif // EXAMPLE_CALCULATE_PI_CALCULATE_PI_HPP_
